Extract FilterHandler from CFFilterHandler
