+++ src/data/c-f-filter-handler.ts
@@ -1,24 +1,12 @@
 import { MinimalCFDimension } from '../core/types';
+import { FilterHandler } from './filter-handler';
 
 export interface ICFFilterHandlerConf {
     dimension?: MinimalCFDimension;
 }
 
-export class CFFilterHandler {
-    private _filters: any[]; // TODO: find better types
-    get filters (): any[] {
-        return this._filters;
-    }
-
-    set filters (value: any[]) {
-        this._filters = value;
-    }
-
+export class CFFilterHandler extends FilterHandler {
     protected _conf: ICFFilterHandlerConf;
-
-    constructor() {
-        this.filters = [];
-    }
 
     public configure(conf: ICFFilterHandlerConf): this {
         this._conf = { ...this._conf, ...conf };
@@ -27,20 +15,6 @@ export class CFFilterHandler {
 
     public conf(): ICFFilterHandlerConf {
         return this._conf;
-    }
-
-    /**
-     * Check whether any active filter or a specific filter is associated with particular chart instance.
-     * This function is **not chainable**.
-     * @see {@link BaseMixin#hasFilterHandler hasFilterHandler}
-     * @param {*} [filter]
-     * @returns {Boolean}
-     */
-    public hasFilter(filter?): boolean {
-        if (filter === null || typeof filter === 'undefined') {
-            return this.filters.length > 0;
-        }
-        return this.filters.some(f => filter <= f && filter >= f);
     }
 
     public applyFilters() {
@@ -71,100 +45,5 @@ export class CFFilterHandler {
                 return false;
             });
         }
-    }
-
-    /**
-     * Filter the chart by the given parameter, or return the current filter if no input parameter
-     * is given.
-     *
-     * The filter parameter can take one of these forms:
-     * * A single value: the value will be toggled (added if it is not present in the current
-     * filters, removed if it is present)
-     * * An array containing a single array of values (`[[value,value,value]]`): each value is
-     * toggled
-     * * When appropriate for the chart, a {@link filters dc filter object} such as
-     *   * {@link filters.RangedFilter `filters.RangedFilter`} for the
-     * {@link CoordinateGridMixin CoordinateGridMixin} charts
-     *   * {@link filters.TwoDimensionalFilter `filters.TwoDimensionalFilter`} for the
-     * {@link HeatMap heat map}
-     *   * {@link filters.RangedTwoDimensionalFilter `filters.RangedTwoDimensionalFilter`}
-     * for the {@link ScatterPlot scatter plot}
-     * * `null`: the filter will be reset using the
-     * {@link BaseMixin#resetFilterHandler resetFilterHandler}
-     *
-     * Note that this is always a toggle (even when it doesn't make sense for the filter type). If
-     * you wish to replace the current filter, either call `chart.filter(null)` first - or it's more
-     * efficient to call {@link BaseMixin#replaceFilter `chart.replaceFilter(filter)`} instead.
-     *
-     * Each toggle is executed by checking if the value is already present using the
-     * {@link BaseMixin#hasFilterHandler hasFilterHandler}; if it is not present, it is added
-     * using the {@link BaseMixin#addFilterHandler addFilterHandler}; if it is already present,
-     * it is removed using the {@link BaseMixin#removeFilterHandler removeFilterHandler}.
-     *
-     * Once the filters array has been updated, the filters are applied to the
-     * crossfilter dimension, using the {@link BaseMixin#filterHandler filterHandler}.
-     *
-     * Once you have set the filters, call {@link BaseMixin#redrawGroup `chart.redrawGroup()`}
-     * (or {@link redrawAll `redrawAll()`}) to redraw the chart's group.
-     * @see {@link BaseMixin#addFilterHandler addFilterHandler}
-     * @see {@link BaseMixin#removeFilterHandler removeFilterHandler}
-     * @see {@link BaseMixin#resetFilterHandler resetFilterHandler}
-     * @see {@link BaseMixin#filterHandler filterHandler}
-     * @example
-     * // filter by a single string
-     * chart.filter('Sunday');
-     * // filter by a single age
-     * chart.filter(18);
-     * // filter by a set of states
-     * chart.filter([['MA', 'TX', 'ND', 'WA']]);
-     * // filter by range -- note the use of filters.RangedFilter, which is different
-     * // from the syntax for filtering a crossfilter dimension directly, dimension.filter([15,20])
-     * chart.filter(filters.RangedFilter(15,20));
-     * @param {*} [filter]
-     * @returns {BaseMixin}
-     */
-    public filter();
-    public filter(filter): this;
-    public filter(filter?) {
-        if (!arguments.length) {
-            return this.filters.length > 0 ? this.filters[0] : null;
-        }
-
-        if (filter === null) {
-            this.resetFilters();
-        } else if (
-            filter instanceof Array &&
-            filter[0] instanceof Array &&
-            !(filter as any).isFiltered
-        ) {
-            // list of filters
-            filter[0].forEach(f => this.toggleFilter(f));
-        } else {
-            this.toggleFilter(filter);
-        }
-
-        this.applyFilters();
-
-        return this;
-    }
-
-    public toggleFilter(filter) {
-        if (this.hasFilter(filter)) {
-            this.removeFilter(filter);
-        } else {
-            this.addFilter(filter);
-        }
-    }
-
-    public addFilter(f) {
-        this.filters.push(f);
-    }
-
-    public removeFilter(filter) {
-        this.filters = this.filters.filter(f => !(filter <= f && filter >= f));
-    }
-
-    public resetFilters() {
-        this.filters = [];
     }
 }
