+++ src/data/index.ts
@@ -2,3 +2,4 @@ export * from './c-f-data-cap-helper';
 export * from './c-f-filter-handler';
 export * from './c-f-multi-adapter';
 export * from './c-f-simple-adapter';
+export * from './filter-handler';
